# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "De'Corrinna API",
        "description": "REST API backend De'Corrinna (Laravel 12 + PostgreSQL). Autentikasi memakai Bearer token (Laravel Sanctum).",
        "contact": {
            "name": "Tim Azatech",
            "email": "dev@azatech.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (dev)"
        }
    ],
    "paths": {
        "/api/admin/addons": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Daftar add-on",
                "operationId": "a712a2743d4d50016e81a7fdf3a0e2d4",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Cari nama add-on.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Aktif",
                                "Nonaktif"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar add-on",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Addon"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak punya permission addons:index",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Tambah add-on",
                "operationId": "aa2dfa834c4871b84bc244f2b7127dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddonInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Add-on dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Addon"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/addons/{addon}": {
            "put": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Ubah add-on",
                "operationId": "6436904675c6cc500a4c515123d2a50e",
                "parameters": [
                    {
                        "name": "addon",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddonInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add-on diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Addon"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Hapus add-on",
                "description": "Soft delete. Tautan ke kategori/item ikut dilepas.",
                "operationId": "40360bdf1f20dc17809547555fc1890d",
                "parameters": [
                    {
                        "name": "addon",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Terhapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Addon": {
                "title": "Addon",
                "description": "Cerminan AddonResource.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Extra Bed"
                    },
                    "price": {
                        "type": "integer",
                        "example": 150000
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "unit": {
                        "description": "Satuan tampilan di A7: \"unit\" (Extra Bed), \"jam\" (Late Checkout), atau null untuk harga flat.",
                        "type": "string",
                        "example": "unit",
                        "nullable": true
                    },
                    "deadline_days": {
                        "description": "Batas pemesanan H-n sebelum check-in (mis. catering H-4).",
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "links": {
                        "description": "Bentuk datar dari pivot polimorfik — mengisi kolom \"Terhubung ke Item/Kategori\".",
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "category",
                                        "item"
                                    ]
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "label": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": [
                            {
                                "type": "category",
                                "id": 1,
                                "label": "Villa De Corrinna"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "AddonInput": {
                "title": "AddonInput",
                "description": "Body untuk POST/PUT add-on. Tautan dikirim sebagai dua daftar id terpisah\nlalu digabung ke pivot polimorfik `addon_links`.",
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Extra Bed"
                    },
                    "price": {
                        "type": "integer",
                        "example": 150000
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "unit": {
                        "type": "string",
                        "example": "unit",
                        "nullable": true
                    },
                    "deadline_days": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    },
                    "item_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": []
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Envelope response error standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan."
                    },
                    "errors": {
                        "description": "Peta field -> daftar pesan (hanya untuk error validasi).",
                        "type": "object",
                        "example": {
                            "email": [
                                "Email wajib diisi."
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Meta pagination standar untuk response terpaginasi.",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Envelope response sukses standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token pribadi dari Laravel Sanctum. Kirim sebagai header: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}